backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.12.0
redis==5.0.1
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-magic==0.4.27
aiofiles==23.2.1
python-docx==1.0.1
html2text==2020.1.16
